+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # 2.x branch
 ## 2.19 branch
 ### 2.19.9
-* added `getAlias()` and `getPluginName()` methods for the `MockTrait`. Fixed
-    `getOriginClassName()` and `getOriginClassNameOrFail()` methods;
+* added `getAlias()`, `getPluginName()` and `getTableClassNameFromAlias()`
+    methods for the `MockTrait`. Fixed `getOriginClassName()` and
+    `getOriginClassNameOrFail()` methods;
 * `MockTrait::getControllerAlias()` method is now deprecated. Use instead `getAlias()`.
 
 ### 2.19.8